+++ app/[locale]/template/[templateId]/section/[section_slug]/page.tsx
@@ -41,7 +41,10 @@ import { DmpEditor } from "@/components/Editor";
 import { SectionFormInterface, SectionFormErrorsInterface, TagsInterface } from '@/app/types';
 import { useSectionData } from "@/hooks/sectionData";
 
+import { useToast } from '@/context/ToastContext';
+
 const SectionUpdatePage: React.FC = () => {
+  const toastState = useToast(); // Access the toast state from context
 
   // Get templateId param
   const params = useParams();
@@ -66,7 +69,6 @@ const SectionUpdatePage: React.FC = () => {
 
   // Save errors in state to display on page
   const [errors, setErrors] = useState<string[]>([]);
-  const [successMessage, setSuccessMessage] = useState<string>('');
   const [fieldErrors, setFieldErrors] = useState<SectionFormInterface>({
     sectionName: '',
     sectionIntroduction: '',
@@ -202,18 +204,14 @@ const SectionUpdatePage: React.FC = () => {
   const handleFormSubmit = async (event: React.FormEvent<HTMLFormElement>) => {
     event.preventDefault();
 
-    setSuccessMessage('');
-
     clearAllFieldErrors();
 
     if (isFormValid()) {
       // Create new section
       await updateSection();
       setErrors([]); // Clear errors on successful submit
-      // For now, scroll to top of page to provide some feedback that form was successfully submitted
-      // TODO: add flash/toast message to signal to user that form was successfully submitted
-      setSuccessMessage(SectionUpdatePage('messages.success'))
-      scrollToTop(topRef);
+      const successMessage = SectionUpdatePage('messages.success');
+      toastState.add(successMessage, { type: 'success', timeout: 3000 });
     }
   };
 
@@ -271,11 +269,6 @@ const SectionUpdatePage: React.FC = () => {
                 </div>
               }
 
-              {successMessage && (
-                <div className="messages success" role="alert" aria-live="assertive">
-                  <p>{successMessage}</p>
-                </div>
-              )}
               <Tabs>
                 <TabList aria-label="Question editing">
                   <Tab id="edit">{Section('tabs.editSection')}</Tab>